move stuff around and fix SPOP
