+++ geode-redis/src/main/java/org/apache/geode/redis/internal/RedisCommandType.java
@@ -15,6 +15,12 @@
 
 package org.apache.geode.redis.internal;
 
+import org.apache.geode.redis.internal.ParameterRequirements.ExactParameterRequirements;
+import org.apache.geode.redis.internal.ParameterRequirements.MaximumParameterRequirements;
+import org.apache.geode.redis.internal.ParameterRequirements.MinimumParameterRequirements;
+import org.apache.geode.redis.internal.ParameterRequirements.ParameterRequirements;
+import org.apache.geode.redis.internal.ParameterRequirements.SpopParameterRequirements;
+import org.apache.geode.redis.internal.ParameterRequirements.UnspecifiedParameterRequirements;
 import org.apache.geode.redis.internal.executor.AuthExecutor;
 import org.apache.geode.redis.internal.executor.DBSizeExecutor;
 import org.apache.geode.redis.internal.executor.DelExecutor;
